[Feature] #29 - admin 대응
